+++ test/configuration_test.rb
@@ -24,6 +24,7 @@ class ConfigurationTest < Minitest::Test
     assert configuration.instrumentation_enabled
     assert_equal "recording_studio_ai", configuration.notification_namespace
     assert_nil configuration.admin_actor_resolver
+    assert_nil configuration.admin_authenticate
     assert_nil configuration.admin_visible_roots_resolver
     assert_nil configuration.admin_layout
     assert_equal 0.1, configuration.admin_warning_thresholds[:error_rate]